Join IPFIX apps together
We may split them apart again, but we need to join them to be able to
nicely generate the flow keys and record types from the templates.
